requirements.txt
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
pip==24.3.1
sqlparse==0.5.3
tzdata==2025.1